Support DLS python3 releases
dls-python3-skeleton is diverging away from the software architecture that the build server expects and the packages that it can provide. Because of this, we require a stopgap that allows us to adopt the skeleton and perform dls releases. This commit adds the Pipfile and setup.py files required for such a release, while also enabling use of dls-python3 (python 3.7). The dependencies in setup.cfg have also been pinned to mach the version of pip used by the build server - 19.3.1. These requirements have been identified and documented here: https://confluence.diamond.ac.uk/display/~hgv27681/Using+python3-pip-skeleton+to+release+into+DLS+Python3